refactor: HighlightGroup
