+++ src/components/HighlightGroup/HighlightGroup.scss
